+++ packages/editor/js/extensions/libs/background/components/index.js
@@ -1,3 +1,5 @@
 // @flow
 export { BackgroundClipping } from './background-clipping';
+export { BackgroundColorField } from './background-color-field';
+export { BackgroundLayersField } from './background-layers-field';
 export { Blending } from './blending';